dual_system: handle the failures of mounting squashfs
Description:
When system bootup, whether ubi attach or mount squashfs
filesystem to fail, the script doesn't handle the error.

Root cause:
In the system startup stage, it doesn't handle to the failure
of ubi attaching and mounting suqashfs.

Solution:
If attach ubi and mount suqashfs to fail, we swap system at
once.

Resolves: QTI9X28-2282 QTI9X40-2598
Change-Id: I505a0d23a3251f8d7576ea85423d9f47630edaaf
Signed-off-by: Brook Zhong <bzhong@sierrawireless.com>