yocto: the layer priority was wrong
When running "bitbake-layers show-layers", will show
an wrong layer priority.

(1) Need to modify "conf/layer.conf" because wrong BBFILE_PATTERN_XXX in
this file is set.
(2) The layer priority of upper directory "meta-swi"
controled the layer priority of lower directory, so move the
"meta-swi/conf" to "common/conf" directory, which was
the same level with "meta-swi-mdm9xxx" and "meta-swi-mdm9x40" directories.

Resolves: QTI9X40-2549 QTI9X28-2309
Change-Id: Ie44927512931781a3b5b8698c633670487e5f43d
Signed-off-by: Lucy He <LHe@sierrawireless.com>