watchdog: Implement watchdog to prevent stuck in kernel startup & shutdown phase
Add three attribute parameters for msm_watchdog, they are "kick",
"stopautokick" and "barktime".
The parameter of "kick" is used to kick MSM_WATCHDOG one time.
The parameter of "stopautokick" is used to stop kicking MSM_WATCHDOG
in kernel.
The parameter of "barktime" is used to set MSM_WATCHDOG barktime in
kernel.

The startup phase:
By default, kernel doesn't kick MSM_WATCHDOG durning bootup
automatically.During system startup, we need to kick WDT when every rcS
script is done to protect module out of stuck. When system is finished
to startup, we need to set barktime back to 11s and start to kick
MSM_WATCHDOG in kernel.

The shutdown phase:
When systme shutdown, set MSM_WATCHDOG barktime to 25s and stop to kick
MSM_WATCHDOG in kernel. Then kick MSM_WATCHDOG one time when every rcK
script is is done.

Resolves: QTI9X28-2367
Change-Id: I2e9a2db9dc6e87f5ae3296e0ab8c9fb5c0263243
Signed-off-by: Lucy He <LHe@sierrawireless.com>