system-core:Adbd(TCP) adopt the same treatment method with AR759x
Issue Description:
adb listen usb port as well as TCP port by default..

Investigation:
In open source, the actions are as below:
If one of the corresponding properties is set, also listen on that port.
If one of the properties isn't set and we couldn't listen on usb,
listen on the default port.

Root Cause:
QC modified the behavior.

Solution:
Revert the commit 8110d90d7d652dc23a1659fed7c89e196f6125b5 in system-core.

Change-Id: I06bf7df3daeaf4f0094f17a6312caf711a4dac56
Resolves: QTI9X28-2314
Signed-off-by: Betty Zhuo <bzhuo@sierrawireless.com>