uart: Remove AT Service initialization on the UART interface
When UART is configured for AT service (modem AT commands), management 
of the interface is done by the modem.  This change disables the AT 
Service initialization on UART which was previously performed in Linux.

Note that the change currently simply comments out the initialization in
this scenario.  Optimization/Clean-up of this change will be done in a 
seperate commit in the future.

Resolves: QTI9X07-330
Change-Id: Iaf02dc8504faf19837ee11f302d6b1d09f67a9f9
Signed-off-by: Kenneth Cheng <kce@sierrawireless.com>
(cherry picked from commit 9d9614c5a06a7aed42f0d388b0e385d02ad62a3f)