mdm9x28/mdm9240: adbd crash issue
Host uses adb to connect module over TCP, then use nmap to scan TCP
port, found adbd crash.

The root cause is the kick function is set to NULL but is used.
This issue does not exist on opensource adbd. It is a bug from Qualcomm
in the commit 53b5c3b9d2170c28a62aef3c329ba6685d4cc25c. 

Revert the code as opensource to fix this issue.

Resolves: QTI9X28-2577 QTI9X40-2507
Change-Id: I02b7daa54dd891edb920cf144ed7974905c7300f
Signed-off-by: Betty Zhuo <bzhuo@sierrawireless.com>