audio: "+WDDI" USL message for DTMF detection is missing
Issue Description:
Enable DTMF decode in 2 modules, make a call on module A,
and send the dtmf to module B, the "+WDDI" USL message
for DTMF detection is missing on module B.

Root Cause:
There is on commit missing to merge on 2.X baseline,
which was related to "Improve the DTMF detect procedure".

Solution:
Merge the corresponding commit from 1.X baseline to 2.X baseline.

Resolves: LXSWIREF-465
Related: QTI9X40-3156
Change-Id: Ib0af82c04f39e8e6512880526a414d5a6d2d8bac
Signed-off-by: kehuang <KeHuang@sierrawireless.com>
Signed-off-by: Christophe Gottheimer <cgottheimer@sierrawireless.com>