toolchain: fix issues in installer "make scripts".
* common/recipes-core/meta/meta-toolchain-swi.bb
(SDK_POST_INSTALL_COMMAND): Fix two problems: wrong variable for root,
and permissions. To cd to the kernel root, we can't expand the
SDKTARGETSYSROOT BitBake variable at build time, but rather must insert
the shell variable $target_sdk verbatim which expands at installer
execution time to the user's chosen target directory. Secondly, "make
scripts" runs into the issue that it can't create files, because
directories in the tree are owned by root. Rather than try to use "sudo"
to do the build (which I tried; there were problems with environment
variable passage and such), what we do is use "sudo" to change the
ownership of directories in the "scripts" subdir of the kernel tree to
match the ID of the installing user. Then run the "make scripts" as that
user, then change the ownership of everything under scripts back to
root.

Resolves: LXSWIREF-305
Change-Id: Ibc07ecb474d672c226206ede38d626835b23b413