mount_early: Can't correctly detach UBI device for legato partition
Description:
During reboot process, when try to detach the UBI device for
legato partition, it missed to check current system number,
but just try to detach it as system1, so when it is in system2
it can't successfully detach the UBI device.

Root cause:
It missed to check the running system number.

Solution:
Check the running system number, and then detach the right
UBI partition device.

Resolves: LXSWIREF-350
Change-Id: I94cf8f48d37b0a2694ce77ba858d81e885371084
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>
Signed-off-by: Duy Doan <dkduy@tma.com.vn>