debug: add lrzsz target package
There are instances where target platform does not have
networking or adb capabilities. However, file transfer
if fairly important part of development cycle, and in some
cases it is nearly impossible to finish the development
cycle without it.

lrzsz package supports file transfers (binary or ascii)
over serial line which is already used as a console. For
more information, please take a look at the
https://ohse.de/uwe/software/lrzsz.html .

Resolves: LXSWIREF-324
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Change-Id: I17124033a0c11c7f9e768ed940aa265d1ce40985