ima: Add IMA handling policy to initramfs
IMA policy must be applied very early in the boot process, and certainly
before any of the "real" file systems are mounted.

Best place to do this is in the initramfs. Initramfs is part of the kernel
image, and that image is verified as part of the chain of trust, while
system is booting up.

This way, policy file is verified, and could be trusted as well.

Resolves: LXSWIREF-332
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Change-Id: I3e83df7a6c87b3dcbcc107bf91f3d251a1becbd5