build: enable build of open-source layers only
Move build script and makefiles to meta-swi so that they could be
used if meta-swi layer is built without meta-swi-extras. When
meta-swi-extras layer is absent, assume that an open-source-only
build is invoked.

Layers meta-mangoh and legato have meta-swi-extras dependencies
to sierra-m2m-init and qmi-framework, respectively. Also, target
machine is determined based on the presence of binary layer files
in meta-swi-extras. In absence of these files, the build has to
be told for which target machine it is being created. This means
that the command for open-source build would have to be:

    MANGOH_BUILD=0 LEGATO_BUILD=0 MACH=<mach> make image_bin

where <mach> is one of mdm9x15, mdm9x28, etc.

Resolves: LXSWIREF-253
Change-Id: I59a2ed0be6c1b667a1d2e730b5b9fca5056b08ac
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>