build: fix open-source build of LK for WP76
When LK source directory is not pulled from apps_proc, it attempts
to pull it from github and from the incorrect branch for WP76.
Also, symlinks to sierra kernel drivers are not created in the
directory pulled from github.

Fix the WP76 branch in github and create a symlink to sierra
drivers in kernel. Add a url to github repo for AR758x.

Resolves: LXSWIREF-337
Change-Id: I70bba98d183402b2232a00ea0844672513ebc5b0
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>