initscripts: create dynamic hostname
In cases where /etc/hostname file is missing or empty, try to
create a hostname from information embedded in the device.

To force this behaviour on target device, rootfs image has to be
built with TARGET_HOSTNAME explicitly set to empty string, e.g.:
    TARGET_HOSTNAME=\"\" make image_bin
Note that quotes in TARGET_HOSTNAME assignment have to be escaped.

Remove hostname setup in syslog's startup script so as not to
override previously assigned hostname.

Resolves: LXSWIREF-260
Change-Id: Ie646397dd900750b5ed7aed2e3119491a2ae6e5e
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>