fakeroot: Fakeroot complains about missing symbols
Starting with libc version 2.24, dlsym() started reporting actual
errors using dlerror(). libc 2.24 is located in our toolchain, and libc
2.23 (which does not report the errors) is part of Ubuntu 16.04
installation. That is why the error is not visible on Ubuntu 16.04, but
it is visible once toolchain environment is setup. fakeroot is always
used from the host machine (note /usr/lib/x86_64-linux-gnu ...) .
In essence, fakeroot is trying to get the address of "acl*" symbols,
fails, and dlsym() in libc 2.24 reports the error. This is valid error,
because these symbols are located in libacl library.

Here, we are:
* Upgrading fakeroot to the latest version with latest patches
  applied.
* Hiding errors reported by libc 2.24, and essentially reverting
  behavior to libc 2.23 .

We do not need fakeroot on the target, and only installing it for
use in SDK.

Patches applied are combination of debian patches from
http://ftp.debian.org/debian/pool/main/f/fakeroot/fakeroot_1.22-2.debian.tar.xz
and Sierra Wireless patches.

Resolves: LXSWIREF-306
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Change-Id: I007c0b4879e1a4fd9eaa5314e07ef5dc4c8ebbcb