ima: Suppress the overlay over /etc if IMA is enabled
For IMA, the overlay over /etc is suppressed. So, the /etc becomes
no more writable. As Legato requires some users and groups to create
apps, /etc/passwd and /etc/group now provide some appLegatoNN entries
reserved for Legato apps.
Check for the presence of "ima_appraise=enforce" on the kernel
command line and remove the /etc overlay if set.

Resolves: LXSWIREF-326
Change-Id: I8c68593d0907477fb89f7816026b6c9b70337c21
Signed-off-by: Christophe Gottheimer <cgottheimer@sierrawireless.com>