mdm9xxx: Enable adbd tools in default image
Re-enable ADBD tools in the default image.  The ADB interface is not
formally supported, but is used in the Sierra development environment for
validation.

Enablement of the ADB USB interface, combined with removal of the
ADBD tools, resulted in a stalled USB startup and a failure to enumerate
on the host.  This change resolves that failure, and the ADB interface
remains disabled by default.

Resolves: QTI9X07-1527
Change-Id: I8a2580f1419fe6b7fc3207f645b07d16a1949e94
Signed-off-by: Ryan Kirk <rkirk@sierrawireless.com>