mdm9xxx: Auto start CMUX daemon
Issue Description:
Add new start script for the CMUX functionality.  This script will
start the DS_MUX daemon with parameter that whether running the
CMUX daemon over UART port or USB port.  Once the script is run,
it will invoke the main function in the DS_MUX daemon and it would
check a file internally and set the CMUX running over UART, or USB,
or disable CMUX and exit the daemon.

Resolves: QTI9X07-452
Change-Id: I221ec78400fd5404818d57b0f89c1aec92bda82b
Signed-off-by: Kenneth Cheng <kce@sierrawireless.com>