mdm9xxx: Improve mount_early and run.env scripts
Global variable "list" in run.env is used to define dual system mtd
partition, but it is defined in mount_early as a local variable to use.
This conflict causes the module cannot reboot normally in some case,
such as the reboot request from dual system.

Need to rename "list" to "DUAL_PARTI_LIST" to avoid this conflict.
Also fix some syntax errors in the scripts.

Resolves: LXSWIREF-595
Related: QTI9X28-3516
Change-Id: I01bca9fd4b87adc50597671be33f6652e848a8ca
Signed-off-by: Betty Zhuo <bzhuo@sierrawireless.com>
Signed-off-by: Nasif Shafi <nshafi@sierrawireless.com>
----
[nshafi] Removed usage of non-standard "==" operator from script.