ima: remove obsolete kernel command line options
According to IMA documentation, ima_tcb option is obsolete.

In addition, this option would force measurement of the files
which should not be measured. For example, policy file
exempts SQUASHFS file system, but process_measurement() call
is still triggered.

If this option is removed, process_measurement() will be called
only for files from the file systems not exempt by ima policy.

Resolves: LXSWIREF-368
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Change-Id: I5075bdf9f0a3ee0ac6a7a1091a62b278ed937ee3