syslog: uses /mnt/legato to detect SMACK disabled mode
As of now the syslog start-up script uses /legato to detect
if legato has SMACK disabled, otherwise it sets the syslog daemon
label to syslog.

Problem is that the /legato mountpoint is only mounted later, when
legato starts. This gives no chance for syslog to be actually able
to detect the mode, and therefore SMACK label is always sets.

This makes the start-up script use /mnt/flash/legato and /mnt/legato
instead of /legato to detect the mode in the case where /legato is not
yet available.

Resolves: LXSWIREF-511
Change-Id: Ic7763180f8d2eaf1759c496615f772a445c7c5e9