gadget: disable adb gadget if adbd is not installed
Gadget implementation for some kernels failed to enumerate if adb
function was enabled and adbd service was not launched. Check for
presence of /sbin/adbd on root filesystem and if absent, remove
ffs (or adb for older kernels) from the list of gadget functions.

Modified original patch, which had been applied against 9x15 and
9x28 but not the AR578x variant.

Resolves: LXSWIREF-979
Related: LXSWIREF-396
Change-Id: Ia9eeb110b35f8426559f6662d66d90e792187b78
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>
Modified-by: Steve MacDonald <smacdonald@sierrawireless.com>
Signed-off-by: Steve MacDonald <smacdonald@sierrawireless.com>