common: fix update-ld-cache detection of stale mounts
The syntax that update-ld-cache relies on to parse 'mount' result is
busybox-specific, resulting in a different behavior on virt since
util-linux is being used as of today.

However, even if mount busybox is used, it appears the the detection of
stale mounts based on 'mount' showing ... (deleted) does not always
works.

This compares that mountpoints are in sync using the inode number,
which should be more reliable, and it uses /proc/mounts instead of
'mount' so as to not rely on a particular 'mount' format.

Resolves: LXSWIREF-602
Change-Id: I5ab1dd2d9917e2934a08869790f2230e920e4648