virt: add kernel headers to SDK
This is to make it possible for Legato to build kernel modules
through the toolchain.

Few things could be inherited on the mdm9xxx side so I factorized
that in the common include
    meta-swi-mdm9xxx/recipes-kernel/linux-quic/linux-quic.inc

The method to generate the kernel-dev package (kernel content that
ends up in the toolchain) was missing some content.
Since Kbuild files were not included, the kernel would remove the
generated headers when running 'make scripts' during the toolchain
install.

Resolves: LXSWIREF-603
Change-Id: I06a8f16b71f7e484d91ce6754d569b969017b854