sdk: fix SDK install as root
Currently when executing the SDK as root,
the SDK already has access to the target directory so SUDO_EXEC
is empty, resulting in an execution issue since it tries to execute
PATH="$PATH".

Instead of sourcing the environment script and then trying to pass
the env. to 'make scripts', this does the sudo (or not) first and
then takes care of sourcing the environment.

Resolves: LXSWIREF-622
Change-Id: I59aa9507cdf0a6c59c06b55f87edd42a0c92686c