common: remove gcc 6.2
gcc has been included in meta-swi as a workaround for
Legato not being compatible with gcc needing the --sysroot option
to be provided (as the default hardcoded sysroot disappeared in
Yocto 1.8 and was replaced by '/not/built').

Since LE-8640 we should be able to resume using the gcc as provided
by poky.

Resolves: LXSWIREF-236
Related: LXSWIREF-237 LE-8640
Change-Id: I3aa83aa824755a060a3b576323e4b86e8b41b2d0