popt, opgk, prelink: pin glob to GLIBC_2.2.5 ver.
Yocto 2.2 is wrongly linking native libraries and executables against
the build machine glibc, which causes them to pick up symbol version
references from the build machine.

For instance, on Ubuntu 18.04 LTS, glibc is 2.27.  A number of functions
in glibc 2.27 are specific to that version (for newly built programs).

One of the functions in this category is the POSIX glob function. Due to
some compatibility issue, it has been bumped to a new version in glibc
2.27.

A few native programs and libraries used during the build process are
calling glob.

When these programs and libraries are executed by various recipes they
fail to load, because at load time, Yocto arranges for them to use its
own prepackaged "sysroot-uninative" glibc that is version 2.25.90.
They consequently cannot load, because that library doesn't offer
the GLIBC_2.27 version of the glob function.

The affected components are libpopt.so (used by debugedit from
rpm-native), the opkg tool (whose libopgk.so uses glob) and also the
prelink utility (uses glob directly).

In this patch, we make these components select the GLIBC_2.2.5 verison
of the glob function, allowing them to link against Yocto's uninative
glibc even when everything is built on a machine running glibc 2.27.

GLIBC_2.2.5 is specified because that's the highest version for the glob
symbol that is offered by Yocto's glibc-2.25.90.

Resolves: LXSWIREF-694
Change-Id: I5d25ca4c28a41373b91d6579f8ddd96d869715bc