common: skip .patch files from legato sources
Some .patch files have been added in the Legato source tree, that
are managed directly by the Legato build system and should not be
taken care of by Yocto.

Yet since they end in .patch, Yocto thinks that it is meant to handle
it and the do_patch task tries to apply them, resulting in a broken
build.

By skipping the .patch files from the sources list, Yocto is not going
to be able to detect changes on these files. This is a relatively small
drawback.

Resolves: LXSWIREF-708
Change-Id: I4316e4790241decefd3f917be55cf840dd87a325