libarchive: stop ignoring Yocto CFLAGS.
The Poky recipe for libarchive wants to pass some
extra flags to the build. See this line:

   CPPFLAGS += "-I${WORKDIR}/extra-includes"

This is so that libarchive can find header files that
were copied into its build directory by do_configure.

However, this does not work at all. All compile steps in libarchive are
handled by libtool and not by a compiler command. It doesn't pick up any
of the Yocto CFLAGS. Not -DSIERRA, none of the code generation
options, nothing!

We can compensate for this in our existing .bbappend for libarchive.

The above CPPFLAGS variable seems to be bogus, anyway;
we want to extend TARGET_CFLAGS. However, the TARGET_CFLAGS are not
being propagated at all.

* common/recipes-extended/libarchive/libarchive_%.bbappend
(TARGET_CFLAGS): Add the -I to reference the extra-includes directory,
as the base recipe requires in order for libarchive to find the ext2
headers which are copied there.
(EXTRA_OEMAKE): Define this variable to pass CFLAGS to the make job.
The libtool command lines issued by the libarchive make *do* interpolate
the $(CFLAGS) variable; we can set CFLAGS to contain the full Yocto
${TARGET_CFLAGS}.

Resolves: LXSWIREF-698
Change-Id: Ia7be4d50a860a4df05c7fbcb482cab6d3834d995