common: remove mtd-utils from layer
These recipes were added a long time ago when we needed something
more recent than what was provided in the Yocto 1.3 poky.

However there is no need for us to keep maintaining that package
as we could get a more recent version (1.5.2 on Yocto 2.2.3 and
2.0.1 on Yocto 2.5).

A tool called 'ubiblkvol' was initially added through a patch, which
is not part of the official mtd-utils. Using 'ubiblock' instead.

Resolves: LXSWIREF-740
Change-Id: Ieb14760cc246d2600998cbce8ecd174e1b1ecbc8