virt: fix dummy-kernel-mod build
The build of extract-cert fails since it relies on the openssl headers.
Unless STAGING_INCDIR_NATIVE and STAGING_LIBDIR_NATIVE are properly
passed, the kernel takes the host lib instead of the Yocto host sysroot
lib.

This was already fixed for the linux-yocto build as part of the IMA on
virt implementation, in file:
 meta-swi/meta-swi-virt/recipes-kernel/linux/linux-yocto_%.bbappend

It appears that in the module.bbclass, it also does a 'make scripts'
in the kernel directory, which also requires that fix.

Resolves: LXSWIREF-742
Change-Id: I128c194bdb2caea926ddb443c8c51bb2b38ff797