mdm9x15: Restore the original wpx5xx-specific restartNMEA script
In the transition to Yocto 2.2 the restartNMEA was centralized for multiple
projects in mdm9xxx.  mdm9x15-based WP variants require a modified script
to bind smd interfaces to support NMEA over USB.  This change restores the
original restartNMEA script for mdm9x15, and fixes the relative priority of
the mdm9x15 layer, above the generic mdm9xxx layer.

Resolves: QTI9X1555-1272
Change-Id: Ia8cb396314fd38f8b644eeea7d1f9f99c7840dd2