virt: deprecate 'virt' legato target in favor of 'virt-x86'/'virt-arm'
The previous way used in a 'LEGATO_VIRT_ARCH' env var to control the arch
of the platform.
This way is a bit more explicit and does not rely on any external conf
but the name of the make target.

The build system is updated to detect support for the old target name,
otherwise it uses the new one.

Resolves: LXSWIREF-782
Change-Id: I2af62fbdfff12150cde962b0d57109b516091a9c