common: remove quota-tools
Licensing of this package is quite strange (some content is GPLv2, and
some is BSD).
Since this is not used at the moment (at least not by Legato), removing
these tools.

We are leaving quota constructs alone, because there may not be
the way to execute quotacheck without utilities. We could potentially
get these from FreeBSD distro (maybe we need to modify them).

We are also leaving up to the customers to put original quota tools
into their distros, and it is much easier to do that when all
quota related constructs are already in.

It still tries to mount the file systems with usrquota,grpquota option
so as to make it possible to control disk quotas programmatically.

Resolves: LXSWIREF-812
Change-Id: Ia4468b8ebd922dc31e89d6ba5a68e2f0ff1854ad
Signed-off-by: Bertrand Roussel <broussel@sierrawireless.com>
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>