linux-quic: use rsync, not cpio for install.
The cpio utility seems to suffer from strange bugs when the tree being
archived/dearchived contains files that are multiply hard linked. This
is now the case because of a "cp -drl" step that is performed by the
do_unpack task to get the kernel tree into the staging area.

If we replace cpio with rsync, the problem goes away.

When extracting the archive, cpio creates some files that have the wrong
contents: they are created as incorrect hard links to some existing
files that have already been extracted.  For instance a shell script in
scripts/ might actually be a hard link to the same object as some
Kconfig file elsewhere.

In 2003, a user reported(*) a very similar behavior to the GNU cpio bugs
mailing list in 2003; nobody replied. It could be the same issue.

---
* http://lists.endsoftwarepatents.org/archive/html/bug-cpio/2003-12/msg00008.html

* common/classes/kernel-src-install.bbclass (kernel_src_install):
Replace "cpio -o | cpio -id" pipe job with a single rsync.
rsync can take a list of paths on its standard input, similarly
to cpio, so the rest of our logic (gathering the list of files
to be copied with find) stays the same.

Resolves: LXSWIREF-811
Change-Id: Ib279befb2fad11514c1597b83d5c6c59bc7d0ba6