common: fix legato-image recipe deps
With Yocto 2.5, the sysroot directories are not global anymore,
but built based on dependencies for each packages.

With legato-image, there is an issue with the legato-tools that
is not populated properly in the native sysroot.

There is also a missing dependency on cwetool-native, fakeroot-native
and cryptsetup-native in some cases.

Also updating virt to depend on legato-image so as to deliver images
in deploy/images/ .

Resolves: LXSWIREF-1110
Related: LXSWIREF-855 LXSWIREF-981
Change-Id: I992a8f1fe11f9c9426e06b9c5ed3f6091202165b
(cherry picked from commit 5c97d55f07f7d4af3b890fdcc7f4c1f82f9ba824)