common: use hash instead of the tag in keyutils recipe
The keyutils_1.5.10.bb recipes is using a tag to fetch the source from
the git repository. If the package is already in download directory,
it will still access internet to check the tag.

If the machine can't access to internet, build fails.

Use hash value to specify version for keyutils.

Modified to use HTTPS and for keyutils.inc.

Resolves: LXSWIREF-1049
Related: QTI9X40-4631
Change-Id: I41106de69d3309d21117a8b7c87270ffb1a48f0e
Signed-off-by: Betty Zhuo <bzhuo@sierrawireless.com>
Signed-off-by: Bertrand Roussel <broussel@sierrawireless.com>
Modified-by: Bertrand Roussel <broussel@sierrawireless.com>