mdm9xxx: fix veritysetup failure in CI env
veritysetup format somehow fails in our CI environment.

It is unsure why, but the detection for direct-io through the
call to open(<hash>, O_DIRECT | ...) passes in the device_ready
function while it fails in device_open.
That is that case only when executed through Yocto.

Direct calls to veritysetup result in operations without direct-io
as the above test in device_ready fails right away.

This change adds a patch to retry without direct-io in device_open.

Note that the patch is only applied for host operations, and should
not affect the behavior on target.

Resolves: LXSWIREF-981 LXSWIREF-982
Change-Id: Ia59a7abb48e02cf7e7540783e10b80f3b21db92e