downloads: use external caches
Instead of using a big git repository that contains lots of tarballs,
this change would rely on external & internal CDN services to provided
the content.

If a downloads/ directory is present at the root, it is still being
used as top priority.
The idea is to ship downloaded content with the Yocto source tarball
so as to make it possible for people to build Yocto without internet
access.

Also providing a way to reuse an internal sstate cache (disabled by
default, and that can be enabled through SHARED_SSTATE=true env var.

Resolves: LXSWIREF-544
Change-Id: Ie84a6442fef2d6828c8a5f3ed979223b47d71447