tiwifi: Defer toggling of WLAN_EN to wl18xx driver
According to:
http://e2e.ti.com/support/wireless-connectivity/wifi/f/968/t/181835
WLAN_EN pin needs to be toggled every time internal ECPU needs a
restart, which is the case across suspend/resume cycles. The driver
needs ownership of WLAN_EN pin (IoT0 GPIO4) to perform this rest.

Relinquish ownership of WLAN_EN pin to wl18xx driver and do not
toggle it in startup script.

Resolves: LXSWIREF-1056
Change-Id: I536dff7a8d1ce507c98fa26d12191dddf1e78934
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>