common: support busybox 1.29.2
When building from Yocto 2.6, it uses busybox 1.29.2 and few patches
do not apply.

Also removing HAVE_RPC and MOUNT_NFS busybox options since this is
not required anymore according to
https://git.yoctoproject.org/cgit/cgit.cgi/poky/commit/?id=87de4a1cc3

Marking CONFIG_TC has not set as to prevent the default option from
conflicting with the iproute2-tc explicitly included in the image.

Resolves: LXSWIREF-1112
Change-Id: I0f03593c19d14cf0b3fd072ebf2e11180c7320c5