common: execute dropbear start-up as admin (and in background)
At the moment the dropbear init sequence runs with the same
SMACK label as busybox, ie _.

This results in /etc/dropbear being admin because it is initially
created by dropbear, which causes issues on the next boot when the
script (_) tries to access /etc/dropbear (admin).

As busybox runs scripts from /etc/init.d by sourcing them,
applying a SMACK64EXEC on /etc/init.d/dropbear has no effect.

Instead, adding a wrapper script that executes /etc/init.d/dropbear,
which now runs with the 'admin' label, and while we are at it we
get it to run in background so as to not block the start-up, especially
during the first boot where the key generation occurs.

Resolves: LXSWIREF-1122
Change-Id: Ia95369eb57b892cdcd4efcdd2e3bb5f720c62409