common: fix overlay detection in dropbear
In change 611f9c40 ("ima: Suppress the overlay over /etc if IMA is enabled"),
a logic is introduced to detected if a overlayfs is mounted in
/etc/init.d/dropbear.

However, since there is a 'set -e' at the top of this file, any error
results in the script exiting, so the logic doesn't work.

Massaging the logic as to make the detection work.

Resolves: LXSWIREF-1154
Related: LXSWIREF-1130
Change-Id: I7360d9210f540069d9a3a49c1609acb8a88624a1
Signed-off-by: Bertrand Roussel <broussel@sierrawireless.com>