smack: add system related rules using system services
Some of the SMACK framework rules must be added very early
in the boot process in order to be available to all other
processes starting up (including Legato).

Legato should not duplicate these rules, because it takes
time to load the rules, and if there are many of them,
timing of the boot process would suffer.

System services will mount smack pseudo file system on standard
mount point: /sys/fs/smackfs .

In order to be backwards compatible, Legato should check if
/etc/smack/accesses file exists. If it does, system services
would be taking care of whatever is inside, and will be mounting
smackfs on /sys/fs/smackfs . If "accesses" file is missing,
Legato should fall back to its legacy behavior.

Change-Id: Id86af403b1598f5b7f95def10d12ee51746eb710
Resolves: LXSWIREF-1047
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>