mdm9607-wp76xx: Fix NMEA and Linux App baudrate at 115200 bps on UART2
Currently, UART 2 is only running on low speed UART driver for NMEA,
and high speed UART driver when Linux APP is used.  Both are running
at baudrate 9600.  But we need to change the baudrate to 115200 bps.

On UART 2, when NMEA is assigned, it is assigned to the low speed
UART driver, and the run.env.in script only assigned to use low
speed UART driver with 9600 bps only from the kernel code.

- In the kernel source code, change the baudrate from 9600 to 115200
- In sierra_serial.c file, assign high speed UART driver when NMEA
  is used.
- In the run.env.in, add new high speed UART driver for UART2, so
  that Linux App and NMEA can be used with high speed UART driver.

Resolves: QTI9X07-2061
Change-Id: I1748e85c40b784cced66b3da588647ca6a1d615c
Signed-off-by: Kenneth Cheng <kce@sierrawireless.com>