bsdiff: package tools as mini shared libs.
We create a libbsdiff.so and libbspatch.so.

* common/recipes-devtools/bsdiff/bsdiff_git.bb (autotools): Do not
inherit this; it is not an autotools project.
(B): Set B same as S; this software doesn't support building in a
separate build dirctory.
(SRC_URI): Add new patch.
(do_patch_extra): Removed. We will not apply the debian patches that are
present in the bsdiff source tree. They contain a flaw that causes a
mmap memory leak when bsdiff is used as a shared lib. Even if that flaw
is fixed, the mmap-based bsdiff seems to performs worse compared to the
malloc+read when used as a shared library, in our delta update use case.
(do_install): Install libbsdiff.so and libbspatch.so.

* common/recipes-devtools/bsdiff/files/0002-package-as-library.patch:
New patch file. (File's header contains change comments.)

Resolves: LXSWIREF-987
Change-Id: I72b336bfee7cc35e3a00a2beb3a1ed4f99eb0b47