initramfs: "garbage collect" unused binaries.
This patch adds a Python script which determines what binaries and
libraries are reachable from the initramfs "./init" script.

We delete binaries and libraries which are not in this set.

The script works by scanning files for strings that are plausible
program and library filenames.  Strings that look like names of other
files are considered to be edges in a graph which is traversed to find
its transitive closure.

* common/classes/swi-image-initramfs.bbclass (remove_entity): Use
argument instead of referring to $file variable. Check for existence;
the caller might now pass us paths that don't exist.
(do_filter_rootfs): Call the gcfs.py script to get a list of paths, and
remove these with xargs rm. Condense the pruning to just do the items
that are not caught by garbage collection. Don't use find to walk the
filesystem; just refer to specific paths.

* common/conf/layer.conf (META_SWI_SCRIPTS): New variable, derived from
LAYERDIR. This points to our new script directory, so we can refer to
scripts.

* common/scripts/gcfs.py: New file (original work).

Resolves: LXSWIREF-1213
Change-Id: I4f15e66bf3429b2c31353cdda030d5711f50818d