wl18xx: build under Yocto 2.6.
* common/recipes-bsp/wl18xx-firmware/wl18xx-firmware_r8.a9.14.bb
(PACKAGE_ARCH): Set value to "${MACHINE_ARCH}" instead of "all".
This firmwware package is pre-built; it certainly doesn't provide
materials for all architectures. Using a value of "all" causes a
problem at image creation time: Yocto thinks that "all" means that a
native version of the package is to be provided also, and then fails
because there is no manifest for such a thing.

Resolves: LXSWIREF-1291
Change-Id: I5a5b4b807bd517b742ce7f945efa11236ce59816