yocto: tools-profile: remove perf & lttng-modules
The perf package in the kernel 3.18 tools/perf directory is defying
attempts to build it. It's quite a mess. Turns out, this is just a
recommend for the tools-profile group. Let's remove the recommendation,
so that enabling tools-profile isn't blocked by this.

The lttng-modules package is also part of tools-profile. It has quite
recently been added in Yocto, and targets newer kernels.

common/recipes-core/packagegroups/packagegroup-core-tools-profile.bbappend: New file.
(RRECOMMENDS): Remove "perf" from the recommended-only packages
for this package.
(RDEPENDS): Remove "lttng-modules".

Resolves: LXSWIREF-1375
Change-Id: I09ccba9af2c6993f5dddc5b9a1d7de10fa9fd1a1