lk: use fetch model so patching works.
* external.mk, repo.mk: Three Yocto variables are now established for
lk. In addition to LK_REPO, there is LK_REPO_DIR and LK_REPO_NAME.
LK_REPO is still a URI, but now it can be relative. The default value
is file://lk.  LK_REPO_DIR is the directory in which this URI is
found, and LK_REPO_NAME is the regular, non-URI name of the repo
within LK_REPO_DIR. The main reason for doing things this way is so that
we don't have to touch the mdm9x15 bbappend for lk, which doesn't
use localgit, and relies on overriding LK_REPO to an external URI.

* meta-swi-mdm9xxx/recipes-bsp/lk/lk_1.3.0.bb (LK_REPO): Do not
establish default value for this variable.
(LK_REPO_DIR): Default this new variable in a way that is safe,
ending up just adding ${THISDIR} to the FILESPATH.
(LK_REPO_NAME): Default this new variable to "git". This works
for recipes like the mdm9x15 that fetch an external git repo,
ending up with a "git" subdirectory in ${WORKDIR}.
(FILESPATH): Prepend ${LK_REPO_DIR}, so we can find the
${LK_REPO} that is expressed with a relative file:// URI.
(S): Don't use "git" as the working directory name; we need a real
workdir, which we obtain from ${LK_REPO_NAME} which gives
the subdirectory that matches the ${LK_REPO} fetch URI.
The SRC_URI stays exactly the same.

* meta-swi-mdm9x40/recipes-bsp/lk/lk_1.3.0.bbappend (localgit):
Inheritance from localgit eliminated, so the regular fetch model is
used.
(SRC_URI): Override to empty string removed; we need the base recipe
SRC_URI to do the fetch; the base recipe locates lk for us.
(SRC_DIR): localgit-specific variable removed.

* meta-swi-mdm9x28/recipes-bsp/lk/lk_1.3.0.bbappend (localgit):
Same changes as for mdm9x40, plus the following:
(do_patch): Symlinking action moved to do_configure_prepend.
A small change is made here: we always create the symlink, deleting
the existing one if necessary. People who pick up this change will
otherwise experience a build problem, due to the symlink existing in
their git repo, and being picked up by fetch.

Resolves: LXSWIREF-1277
Change-Id: I83897907fb1880235c4c6382e50bb289eb85e8da