sdx55: Support DM verity
Description:
The purpose of dm-verity is to implement a device mapper target
capable of validating the data blocks contained in a file system
against a list of cryptographic hash values.

We can close this function by set it to "DM_VERITY_ENCRYPT = off"
in file "swi-sdx55.conf". The root hash of the rootfs is stored
in scripts in initramfs, So when dm-verity is enabled we must burn
the kernel and the rootfs together, they are pairs or you will got
crash during boot up.

Dm-verity is actived from initramfs. Till now only the rootfs partition
is protected by Dm-verity.The rootfs data is stored at UBI volume 0 (UBI0_0)
and the corresponding data hash tree table is store at volume 1 (UBI0_1)
Every access to the data blocks (mtdblock0_0) Dm-verity will check
corrsponding hash tree table (mtdblock0_1). If the data and the hash is
not match then return I/O error, which is expected.

Solution:
Porting DM verity feature from mdm9x40.
Use libgcrypt to support DM verity instead of openssl.

Resolves: LXSWIX55-37
Change-Id: If2538f56c21d5843d68029ca1ec878984b795c83
Signed-off-by: Betty Zhuo <bzhuo@sierrawireless.com>