wifi: Co-Existence for BX321x IoT is not working
The LTE-WiFi Coexistence with BX321x IoT card is not working. The UART pin for CoEx
does not output any signal and there is no communication between mdm and
wlan module to prevent the channel conflict.

Automatic Channel Selection (ACS) has to be enabled for WLAN channel avoidance
WCN policy. The following binary files are required to be placed in the correct
wlan dirver path for enabling LTE Coex:
- bdwlan.bin : Given by Qualcomm in SR 04101508 with LTE Coex enabled.
- qwlan.bin : The same file given in our source code
              (mdm9x28/wlan_firmware/wlan/fw/target/.output/AR6320/hw.3/bin/qwlan.bin)

The coex_config_data NV has to be updated too but it is out of the scope of this
commit.

Resolves: QTI9X07-3281
Change-Id: I1adbe41e1b6968336ec123a970160cd382ab4715
Signed-off-by: Alan Cheung <alcheung@sierrawireless.com>