nagger: Rework nagger for generic mdm9xxx and fx30
Undo part of the previous commit a10e7541a3faf271. That commit makes
the nagger run as login shell, which introduces a rollback issue.
This commit undoes that but keeps the new nagger script (which contains
some commonly desired features).

Both generic mdm9xxx (eg. wp76/77, ar75xx) and fx30 project now shares a
same nagger script, and it resides in /usr/sbin. For generic mdm9xxx,
it is invoked by /bin/sh AFTER login with a symbolic link inside
/etc/profile.d. For fx30, it is invoked as the login shell and the symbolic
link is not present.

In order to resolve the roll-back issue on generic mdm9xxx, we decided to:
1) revert root's login shell to be /bin/sh (rolling back to an older version
will not trigger the file-not-found error)
2) keep the nagger script in /usr/sbin/ (enforces better security)
3) make a symbolic link to /usr/sbin/loginNagger in /etc/profile.d/ (so that
loginNagger can be invoked after 1) and 2) is performed)

Resolves: LXSWIREF-1510
Change-Id: I3eb33d8f108447536f5620031eea83e0d4747f78
Signed-off-by: Kevin Liu <keliu@sierrawireless.com>