common: support nohup when running through USE_DOCKER=1
When running as:
  USE_DOCKER=1 nohup make image

it currently tries to launch the Docker container with
--tty --interactive (= -ti)
This doesn't work since the shell is not a tty.
Adding a detection to only use this option in the case
where the shell is a tty.

Also using --init to start the process using a init
that is going to reap processes.

Resolves: LXSWIREF-1580
Change-Id: I390470ad827e97ae0743f3ed72d5af87d2c498ae