nagger: Login nagger should work with IMA enabled
Once IMA is enabled, /etc will not be part of overlay file system.
In that case, login nagger would not be able to change user password,
and dropbear keys would be always generated on the fly.

In order to overcome these problems, we would need to allow password
utility as well as dropbear and login nagger to store its required
content onto persistent storage.

Note that device could be exposed to the attacks if root password is
not set at build time, and both /etc and /mnt/flash are not
writable.

Change-Id: Ia82ee1df3586226055882254f145726bee62263a
Resolves: LXSWIREF-1579
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>