userland, shadow: Do not use real lckpwdf()
shadow utility is using lckpwdf() method from libc. This
represents the problem, because lckpwdf() is storing
/etc/shadow related lock file in /etc, and if /etc is not
writable lock could not be obtain and password change
operation will fail.

The code which does not depend on lckpwdf(), stores
lock file at the location of shadow file. This approach
is much more suitable in cases /etc/ is not writable.

Resolves: LXSWIREF-1617
Related: LXSWIREF-1579
Change-Id: Iad82b11fac541dcfeae220925269053f9d5a8e39
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>